Turn locking functions in apc_sma.c into a set of locking macros to make it easier to replace the locking mechanism for people who need that.
Benchmarked blocking interruptions in the sma locking macros, but the
performance hit was too big, so blocked at a higher level in apc_main.c
instead around the major shared memory copy functions.

Finally, support the case where the we don't get a filepath passed in
via the zend_file_handle.  In this case pick the filepath out of the
cache entry.