Remove TSRMLS_FETCH() calls inside our error functions, this patch also renames the functions to their respective function rather than "apc_<x>printf(...)". Below is a shows the renamed functions:
 - apc_eprint - apc_error
 - apc_wprint - apc_warning
 - apc_nprint - apc_notice
 - apc_dprint - apc_debug

Notice that the apc_debug() function now only is available if the 
__DEBUG_APC__ macro is definfed.

Error call signature have changed aswell and required the TSRMLS 
macros to be passed at every call. Below example illustrates whats 
change:

 Before:
 apc_wprint("Unable to allocate %d bytes", bytes);

 After:
 apc_warning("Unable to allocate %d bytes" TSRMLS_CC, bytes));

Also Thanks to Pierrick (adoy@) for testing and helping with a few 
fixes for the UNIX build.